+++ tests/test_analyzers.py
@@ -5,7 +5,7 @@ Test TBsim analyzers
 import numpy as np
 import sciris as sc
 import starsim as ss
-import starsim.library as ssl
+import starsim.library.networks as ssln
 import tbsim
 import pytest
 
@@ -29,7 +29,7 @@ def test_householdstats():
     hh_ids = np.arange(50)
     ages = [sc.strjoin(np.random.randint(1, 70, np.random.randint(2, 6))) for _ in hh_ids]
     dhs = sc.dataframe(hh_id=hh_ids, ages=ages)
-    net = ssl.networks.HouseholdNet(dhs_data=dhs, dynamic=False)
+    net = ssln.HouseholdNet(dhs_data=dhs, dynamic=False)
     az = tbsim.HouseholdStats(save_at=['2001-01-01', '2004-01-01'])
     sim = make_sim(analyzers=az, networks=net, copy_inputs=False)
     sim.run()